requirements.txt
streamlit
pandas
numpy
tensorflow
python-binance
requests
streamlit-autorefresh
plotly
nltk
scikit-learn